Refactor: 온보딩 Days Cell 코드 2개 파일 분리
